% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combustion_equation.R
\name{get_combustion_equation}
\alias{get_combustion_equation}
\title{get_combustion_equation}
\usage{
get_combustion_equation(molecule)
}
\arguments{
\item{molecule}{A character vector with one element, to describe a chemical
formula}
}
\value{
A tibble with the balanced combustion equation
}
\description{
get_combustion_equation
}
\examples{
x <- "CH4"
get_combustion_equation(x)
}
