% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements.R
\name{get_elements}
\alias{get_elements}
\title{Convert a chemical molecule into its constituent elements with its respective counts as a dictionary.}
\usage{
get_elements(molecule)
}
\arguments{
\item{molecule}{string}
}
\value{
dictionary
}
\description{
Convert a chemical molecule into its constituent elements with its respective counts as a dictionary.
}
\examples{
get_elements("(C2H4)5")
get_elements("Al2(SO4)3")
}
