% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_molec_props.R
\name{get_molec_props}
\alias{get_molec_props}
\title{get_molec_props}
\usage{
get_molec_props(molecule)
}
\arguments{
\item{molecule}{A character vector with one molecule}
}
\value{
A dataframe with various atomic properties of the elements in the
molecule
}
\description{
get_molec_props
}
\examples{
get_molec_props('CH4')

}
