% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid}
\alias{is_valid}
\title{Check whether a given molecule in string format is chemically valid}
\usage{
is_valid(molecule, simple_deck = FALSE)
}
\arguments{
\item{molecule}{logical}
}
\value{
logical
}
\description{
Check whether a given molecule in string format is chemically valid
}
\examples{
is_valid("NaOH")
is_valid("HF")
is_valid("Al(SO4)3")
}
